Pick an overload for the mobile keep distance
ClientViewRange is a byte and the limit is an int constant, so Math.Min
could not tell its (byte, byte) and (int, int) overloads apart.

+++ src/ClassicUO.Client/Game/World.cs
@@ -92,7 +92,7 @@ namespace ClassicUO.Game
         /// server still describes their movement - see Constants.MAX_MOBILE_VIEW_RANGE.
         /// </summary>
         public static int MobileKeepRange =>
-            Math.Min(ClientViewRange, Constants.MAX_MOBILE_VIEW_RANGE);
+            Math.Min((int)ClientViewRange, Constants.MAX_MOBILE_VIEW_RANGE);
 
         public static bool SkillsRequested { get; set; }
